refactor pool k cache
